move error modal test into upload photos
